feature: json_data argument containing dashboard
path module argument needs to read file on remote host which is
redundant if module is transmitted including data. Therefore json_data
provides a method to use a lookup plugin which runs on the local host
running the ansible playbook to provide dashboard data e.g. via reading
it from a local file or downloading a url locally.